requirements.txt
google-genai
streamlit
pyperclip